+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- The scout:index and scout:delete commands are now implemented for Elasticsearch
 - A DocumentAdapterInterface with an adapter that only deals with documents
 
 ### Changed